+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 2.1.5
+- Git-based package update
+
+# 2.1.4
+- Automatic return URL generation
+
 # 2.1.3
 - Subshop wrong payment method redirect issue fix